Reduced number of max supported maps from 256 to 40 This should optimize some functions related to map sectors/regions
The main idea was reduce to 16, but for some unknown reason, any value < 39 will cause weird map issues like NPCs walking over mountains or static items. After some time searching I didn't found anything that should fix this, so I just set 40 which is working fine

+++ src/network/send.cpp
@@ -3826,8 +3826,8 @@ PacketEnableMapDiffs::PacketEnableMapDiffs(const CClient* target) : PacketExtend
 {
 	ADDTOCALLSTACK("PacketEnableMapDiffs::PacketEnableMapDiffs");
 
-	size_t map = 0;
-	size_t mapCount = 1;
+	int map = 0;
+	int mapCount = 1;
 
 	// Find map count
 	for (map = MAP_QTY - 1; map >= 0; --map)
